style: update eslint config with `@typescript-eslint` and use `next lint` with `lint-staged`
`@typescript-eslint` contains additional linting rules that are not
applied with `extends: ["next/core-web-vitals", "prettier"]` by default.
Specifically, I'm looking to add the
[`no-floating-promises`](https://typescript-eslint.io/rules/no-floating-promises/)
and
[`no-misused-promises`](https://typescript-eslint.io/rules/no-misused-promises/)
rules.

This commit adds these rules (and some more) and corrects the linting
errors this caused.

This commit also fixes our lint-staged config to use the `next lint`
command rather than `eslint` (`eslint` is used under the hood by
`next lint`), following steps from official docs:
https://nextjs.org/docs/pages/building-your-application/configuring/eslint#lint-staged

+++ src/cloudflare/api.ts
@@ -30,12 +30,17 @@ const api = wretch("https://api.cloudflare.com/client/v4/") //
         return next(url, opts);
       },
   ])
-  .auth(`Bearer ${cloudflareApiToken}`);
+  .auth(cloudflareApiToken ? `Bearer ${cloudflareApiToken}` : "");
 
 export const cloudflareApi = {
   images: {
-    createImageUploadUrl: (params: CloudflareDirectUploadParameters) =>
-      api
+    createImageUploadUrl: (params: CloudflareDirectUploadParameters) => {
+      if (!cloudflareAccountId) {
+        throw new Error(
+          "CLOUDFLARE_ACCOUNT_ID is not set. The cloudflare API cannot be used."
+        );
+      }
+      return api
         .url(`accounts/${cloudflareAccountId}/images/v2/direct_upload`)
         .formData({
           ...params,
@@ -45,10 +50,16 @@ export const cloudflareApi = {
           async (r) =>
             await CloudflareDirectUploadDTOSchema.parseAsync(await r.json())
         )
-        .post(),
+        .post();
+    },
 
-    isUploaded: (cloudflareImageId: string): Promise<boolean> =>
-      api
+    isUploaded: (cloudflareImageId: string): Promise<boolean> => {
+      if (!cloudflareAccountId) {
+        throw new Error(
+          "CLOUDFLARE_ACCOUNT_ID is not set. The cloudflare API cannot be used."
+        );
+      }
+      return api
         .url(`accounts/${cloudflareAccountId}/images/v1/${cloudflareImageId}`)
         .resolve((r) =>
           r
@@ -58,15 +69,22 @@ export const cloudflareApi = {
                 !(await CloudflareImageDTOSchema.parseAsync(r)).result.draft
             )
         )
-        .get(),
+        .get();
+    },
 
-    delete: (cloudflareImageId: string) =>
-      api
+    delete: (cloudflareImageId: string) => {
+      if (!cloudflareAccountId) {
+        throw new Error(
+          "CLOUDFLARE_ACCOUNT_ID is not set. The cloudflare API cannot be used."
+        );
+      }
+      return api
         .url(`accounts/${cloudflareAccountId}/images/v1/${cloudflareImageId}`)
         .resolve(
           async (r) =>
             await CloudflareResultDTOSchema.parseAsync(await r.json())
         )
-        .delete(),
+        .delete();
+    },
   },
 };
